Fix audio truncation at end of file when using ALSA.
The problem actually occurs on all audios but is just much more
obvious on those of subsecond duration.

MOC uses ALSA's non-blocking mode and when the audio device is
closed ALSA just discards any unplayed samples.  Turning on
blocking mode and asking ALSA to drain the device should be
the solution to this, but it also triggers bugs in ALSA which
either ignore the setting [1] or cause unacceptable delays [2,3].
There is a report that at least the delays are fixed in alsalib
1.0.25 [4] and we should revisit this patch when we have that
ALSA release installed.

For now, we loop around a usleep(3) while there are still
samples in ALSA's buffers.  However, the method of detecting
when the buffers are drained (via an -EPIPE return code) seems
less than satisfactory.

Resolves: node/817 (Problem playing very short WAV files)
Possibly resolves: node/903 (Broken record syndrome) [5]
Refs: [1] http://ceph.com/git/?p=ceph-client.git;a=commitdiff_plain;h=4cdc115fd38b54642e8536a5c2389483bcb9b2e9
      [2] http://http://code.google.com/p/chromium/issues/detail?id=131136
      [3] http://comments.gmane.org/gmane.linux.alsa.devel/102753
      [4] http://http://code.google.com/p/chromium/issues/detail?id=131136#c4
      [5] http://src.chromium.org/viewvc/chrome?revision=138939&view=revision
Also see: http://listarc.com/showthread.php?3205254-snd_pcm_drain+and+partial+periods
          http://stackoverflow.com/questions/3936420/alsa-how-to-tell-when-a-sound-is-finished-playing