Use different method for ensuring ALSA buffer is drained.
The first approach to solving the ALSA audio truncation problem worked
for some files, but for others it resulted in an endless loop around the
usleep(3).  There does not seem to be any reliable way of knowing when
all the audio samples have been played so the device can be closed.

This approach causes MOC to block until the samples are drained, but it
may introduce unacceptable delays between files because of bugs in ALSA.
If it does, then we may need to provide an option so users can choose
between audio truncation and excessive delays.  Neither problem is one
for which a solution in MOC is obvious from the ALSA documentation.

Amends: r2550
Also see: http://stackoverflow.com/questions/3936420/alsa-how-to-tell-when-a-sound-is-finished-playing

git-svn-id: svn://svn.daper.net/moc/trunk@2553 910807d9-36e0-0310-a014-e9ea483e2ba4