Send error message to clients with server error events.
The server used a static buffer from which the error message had
to be retrieved via CMD_GET_ERROR after an EV_SRV_ERROR was sent.
That meant that only the latest error message was ever seen, and
if there were multiple outstanding errors then that last message
was shown multiple times.

git-svn-id: svn://svn.daper.net/moc/trunk@2649 910807d9-36e0-0310-a014-e9ea483e2ba4