Send error message to clients with server error events.
The server used a static buffer from which the error message had
to be retrieved via CMD_GET_ERROR after an EV_SRV_ERROR was sent.
That meant that only the latest error message was ever seen, and
if there were multiple outstanding errors then that last message
was shown multiple times.