Fix event notification suppression for some client configurations.
If the client built a playlist from the command line arguments and
'SyncPlaylist' was set to 'no', then the client did not send the
server a CMD_SEND_EVENTS command.  Knowing the client did not want
events, the server did not send them when they occurred.

Unfortunately, this suppressed *all* event notifications, not just
those relating to playlists.  Therefore, the client did not know
(and could not display) the current state of the server.