Disable the rendering of the POPT-interpreted command line on OpenWRT.
For reasons which we have not been able to determine, the freeing of
the cloned POPT option table (required to render the POPT-interpreted
command line) causes the assertion on 'opts' in free_popt_clone() to
fire (with NDEBUG undefined) or a segfault (with NDEBUG defined).

The MOC code seems correct and works on every other platform.  It
may be something specific to POPT on OpenWRT or the uClibc library
which it uses.

Until we get additional debugging information we'll simply disable
POPT-interpreted command line rendering (for logging and the output
of the --echo-args option) on OpenWRT.

git-svn-id: svn://svn.daper.net/moc/trunk@2763 910807d9-36e0-0310-a014-e9ea483e2ba4