C99: Silence function pointer conversion warnings.
This looks messy but is the recommended way of doing it.  It
may not work for some now rare hardware architectures, but then
a simple assignment would not have worked either.  It's very
unlikely that MOC would ever be run on such architectures, so
we'll cross that bridge when we come to it.

git-svn-id: svn://svn.daper.net/moc/trunk@2797 910807d9-36e0-0310-a014-e9ea483e2ba4