Fix: Broken pipe on sending EV_EXIT to client originating CMD_QUIT.
At every point where the client sends the server a CMD_QUIT it then
immediately closes the socket connection, so it's pointless the server
attempting to send it an EV_EXIT event (besides, that client already
knows) and it just results in the logging of a "Broken pipe" error
message.  So we have the server tear down that client's connection
immediately.

git-svn-id: svn://svn.daper.net/moc/trunk@2814 910807d9-36e0-0310-a014-e9ea483e2ba4