C99: strcasestr(3) is a GNU extension.
It is exposed by 'string.h' if we define '_GNU_SOURCE', but we're
aiming for POSIX compliance so we don't.  However, the C library
implementation will still be present if we're linking against GNU's
libc so we'll use that if we can (even though using our own would
not cause a conflict).

The C strcasestr(3) declaration also conflicts with that for C++
with regard to constness, so we don't expose it in C++ code.

git-svn-id: svn://svn.daper.net/moc/trunk@2837 910807d9-36e0-0310-a014-e9ea483e2ba4