FFmpeg: Use MOC's I/O functions for files as well as streams.
This simplifies the code and gives MOC more control over the library's
I/O processing.

This change triggers an FFmpeg bug [1] which causes it to leak 1MiB
sized chunks of memory when attempting to open an empty or corrupt
file.  There is no obvious remedy within MOC for this.

It also triggers an FFmpeg bug [2] which requires the W64-format to
be disabled from FFmpeg-1.1 onwards.

Refs: [1] https://trac.ffmpeg.org/ticket/5584
      [2] https://trac.ffmpeg.org/ticket/5610

git-svn-id: svn://svn.daper.net/moc/trunk@2880 910807d9-36e0-0310-a014-e9ea483e2ba4