Fix: Freeze on Nokia devices when audio plays to completion.
This circumvents a freeze of the MOC server when running on a
Nokia N810 tablet (and probably devices with similar soundcards).

If playing is not manually stopped and no other audio files are to
be played when the current audio file reaches the end (either because
the directory/playlist is exhausted or the 'AutoNext' toggle is off)
then the server effectively enters a infinite duration sleep.

The problem arises from an undocumented (and probably erroneous)
negative frame count delay value returned by ALSA's snd_pcm_delay()
function.

Resolves: node/1568 (MOC Freezes on a Nokia N810 Tablet)

git-svn-id: svn://svn.daper.net/moc/trunk@2920 910807d9-36e0-0310-a014-e9ea483e2ba4