Maint: Suppress Clang's overly-enthusiastic variadic macro warnings.
Clang flags as warnings uses of GNU's CPP extension which allows for
the declaration and use of variadic macros having zero variable arguments.
This is controlled by Clang's '-Wgnu-zero-variadic-macro-arguments'
option and also enabled by '-pedantic'.

Unfortunately, this is a very useful extension to have and avoiding
it would result in rather inelegant code.  Fortunately, it can be
suppressed from within MOC.  No one has yet complained about the
several GNUisms remaining in the code, so it can be assumed that those
who are building MOC are doing so with a GCC-compatible compiler.

git-svn-id: svn://svn.daper.net/moc/trunk@2985 910807d9-36e0-0310-a014-e9ea483e2ba4